@@ -1,6 +1,6 @@
 {
   "name": "react-native-gifted-chat",
-  "version": "0.0.10",
+  "version": "0.2.9",
   "description": "The most complete chat UI for React Native",
   "main": "index.js",
   "repository": {
@@ -23,19 +23,28 @@
     "url": "https://github.com/FaridSafi/react-native-gifted-chat/issues"
   },
   "homepage": "https://github.com/FaridSafi/react-native-gifted-chat#readme",
+  "scripts": {
+    "lint": "eslint . --ext .js,.jsx"
+  },
   "devDependencies": {
-    "eslint": "^3.2.2",
-    "eslint-plugin-react": "^6.0.0",
-    "eslint-plugin-react-native": "^2.0.0"
+    "eslint": "3.19.0",
+    "eslint-config-airbnb": "15.0.2",
+    "eslint-config-cooperka": "0.2.3",
+    "eslint-plugin-import": "2.6.1",
+    "eslint-plugin-jsx-a11y": "5.1.1",
+    "eslint-plugin-react": "7.1.0",
+    "eslint-plugin-react-native": "2.3.2"
   },
   "dependencies": {
-    "@exponent/react-native-action-sheet": "git+https://github.com/FaridSafi/react-native-action-sheet.git",
-    "md5": "^2.1.0",
-    "moment": "^2.13.0",
-    "react-native-communications": "^2.1.0",
-    "react-native-dismiss-keyboard": "^1.0.0",
-    "react-native-invertible-scroll-view": "^1.0.0",
-    "react-native-parsed-text": "^0.0.15",
-    "shallowequal": "^0.2.2"
+    "@expo/react-native-action-sheet": "^1.0.1",
+    "md5": "2.2.1",
+    "moment": "^2.19.0",
+    "prop-types": "15.5.10",
+    "react-native-communications": "2.2.1",
+    "react-native-invertible-scroll-view": "^1.1.0",
+    "react-native-lightbox": "^0.7.0",
+    "react-native-parsed-text": "^0.0.19",
+    "shallowequal": "1.0.2",
+    "uuid": "3.1.0"
   }
 }
