@@ -1,11 +1,8 @@
-import React from 'react';
-import {
-  Image,
-  StyleSheet,
-  View,
-} from 'react-native';
-
-import GiftedAvatar from './GiftedAvatar';
+import PropTypes from 'prop-types';
+import React from "react";
+import {Image, StyleSheet, View, ViewPropTypes} from "react-native";
+import GiftedAvatar from "./GiftedAvatar";
+import {isSameUser, isSameDay, warnDeprecated} from "./utils";
 
 export default class Avatar extends React.Component {
   renderAvatar() {
@@ -17,12 +14,21 @@ export default class Avatar extends React.Component {
       <GiftedAvatar
         avatarStyle={StyleSheet.flatten([styles[this.props.position].image, this.props.imageStyle[this.props.position]])}
         user={this.props.currentMessage.user}
+        onPress={() => this.props.onPressAvatar && this.props.onPressAvatar(this.props.currentMessage.user)}
       />
     );
   }
 
   render() {
-    if (this.props.isSameUser(this.props.currentMessage, this.props.nextMessage) && this.props.isSameDay(this.props.currentMessage, this.props.nextMessage)) {
+    const renderAvatarOnTop = this.props.renderAvatarOnTop;
+    const messageToCompare = renderAvatarOnTop ? this.props.previousMessage : this.props.nextMessage;
+    const computedStyle = renderAvatarOnTop ? "onTop" : "onBottom"
+
+    if (this.props.renderAvatar === null) {
+      return null
+    }
+
+    if (isSameUser(this.props.currentMessage, messageToCompare) && isSameDay(this.props.currentMessage, messageToCompare)) {
       return (
         <View style={[styles[this.props.position].container, this.props.containerStyle[this.props.position]]}>
           <GiftedAvatar
@@ -31,8 +37,10 @@ export default class Avatar extends React.Component {
         </View>
       );
     }
+
     return (
-      <View style={[styles[this.props.position].container, this.props.containerStyle[this.props.position]]}>
+      <View
+        style={[styles[this.props.position].container, styles[this.props.position][computedStyle], this.props.containerStyle[this.props.position]]}>
         {this.renderAvatar()}
       </View>
     );
@@ -42,8 +50,12 @@ export default class Avatar extends React.Component {
 const styles = {
   left: StyleSheet.create({
     container: {
-      marginRight: 8,
+      marginRight: 8
     },
+    onTop: {
+      alignSelf: "flex-start"
+    },
+    onBottom: {},
     image: {
       height: 36,
       width: 36,
@@ -54,6 +66,10 @@ const styles = {
     container: {
       marginLeft: 8,
     },
+    onTop: {
+      alignSelf: "flex-start"
+    },
+    onBottom: {},
     image: {
       height: 36,
       width: 36,
@@ -63,8 +79,7 @@ const styles = {
 };
 
 Avatar.defaultProps = {
-  isSameDay: () => {},
-  isSameUser: () => {},
+  renderAvatarOnTop: false,
   position: 'left',
   currentMessage: {
     user: null,
@@ -72,14 +87,26 @@ Avatar.defaultProps = {
   nextMessage: {},
   containerStyle: {},
   imageStyle: {},
+  //TODO: remove in next major release
+  isSameDay: warnDeprecated(isSameDay),
+  isSameUser: warnDeprecated(isSameUser)
 };
 
 Avatar.propTypes = {
-  isSameDay: React.PropTypes.func,
-  isSameUser: React.PropTypes.func,
-  position: React.PropTypes.oneOf(['left', 'right']),
-  currentMessage: React.PropTypes.object,
-  nextMessage: React.PropTypes.object,
-  containerStyle: View.propTypes.style,
-  imageStyle: React.PropTypes.oneOfType([View.propTypes.style, Image.propTypes.style]),
+  renderAvatarOnTop: PropTypes.bool,
+  position: PropTypes.oneOf(['left', 'right']),
+  currentMessage: PropTypes.object,
+  nextMessage: PropTypes.object,
+  onPressAvatar: PropTypes.func,
+  containerStyle: PropTypes.shape({
+    left: ViewPropTypes.style,
+    right: ViewPropTypes.style,
+  }),
+  imageStyle: PropTypes.shape({
+    left: ViewPropTypes.style,
+    right: ViewPropTypes.style,
+  }),
+  //TODO: remove in next major release
+  isSameDay: PropTypes.func,
+  isSameUser: PropTypes.func
 };
