@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import {
   Platform,
@@ -6,21 +7,41 @@ import {
 } from 'react-native';
 
 export default class Composer extends React.Component {
+  onContentSizeChange(e) {
+    const contentSize = e.nativeEvent.contentSize;
+
+    // Support earlier versions of React Native on Android.
+    if (!contentSize) return;
+
+    if (!this.contentSize || this.contentSize.width !== contentSize.width || this.contentSize.height !== contentSize.height) {
+      this.contentSize = contentSize;
+      this.props.onInputSizeChanged(this.contentSize);
+    }
+  }
+
+  onChangeText(text) {
+    this.props.onTextChanged(text);
+  }
+
   render() {
     return (
       <TextInput
         placeholder={this.props.placeholder}
         placeholderTextColor={this.props.placeholderTextColor}
         multiline={this.props.multiline}
-        onChange={(e) => {
-          this.props.onChange(e);
-        }}
-        style={[styles.textInput, this.props.textInputStyle, {
-          height: this.props.composerHeight,
-        }]}
+
+        onChange={(e) => this.onContentSizeChange(e)}
+        onContentSizeChange={(e) => this.onContentSizeChange(e)}
+
+        onChangeText={(text) => this.onChangeText(text)}
+
+        style={[styles.textInput, this.props.textInputStyle, { height: this.props.composerHeight }]}
+
+        autoFocus={this.props.textInputAutoFocus}
+
         value={this.props.text}
         accessibilityLabel={this.props.text || this.props.placeholder}
-        enablesReturnKeyAutomatically={true}
+        enablesReturnKeyAutomatically
         underlineColorAndroid="transparent"
         {...this.props.textInputProps}
       />
@@ -46,7 +67,6 @@ const styles = StyleSheet.create({
 });
 
 Composer.defaultProps = {
-  onChange: () => {},
   composerHeight: Platform.select({
     ios: 33,
     android: 41,
@@ -57,15 +77,22 @@ Composer.defaultProps = {
   textInputProps: null,
   multiline: true,
   textInputStyle: {},
+  textInputAutoFocus: false,
+  onTextChanged: () => {
+  },
+  onInputSizeChanged: () => {
+  },
 };
 
 Composer.propTypes = {
-  onChange: React.PropTypes.func,
-  composerHeight: React.PropTypes.number,
-  text: React.PropTypes.string,
-  placeholder: React.PropTypes.string,
-  placeholderTextColor: React.PropTypes.string,
-  textInputProps: React.PropTypes.object,
-  multiline: React.PropTypes.bool,
+  composerHeight: PropTypes.number,
+  text: PropTypes.string,
+  placeholder: PropTypes.string,
+  placeholderTextColor: PropTypes.string,
+  textInputProps: PropTypes.object,
+  onTextChanged: PropTypes.func,
+  onInputSizeChanged: PropTypes.func,
+  multiline: PropTypes.bool,
   textInputStyle: TextInput.propTypes.style,
+  textInputAutoFocus: PropTypes.bool,
 };
