@@ -1,47 +1,32 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import {
   View,
+  ViewPropTypes,
   StyleSheet,
 } from 'react-native';
 
-import moment from 'moment';
-
 import Avatar from './Avatar';
 import Bubble from './Bubble';
+import SystemMessage from './SystemMessage';
 import Day from './Day';
+
+import {isSameUser, isSameDay} from './utils';
 
 export default class Message extends React.Component {
 
-  isSameDay(currentMessage = {}, diffMessage = {}) {
-    let diff = 0;
-    if (diffMessage.createdAt && currentMessage.createdAt) {
-      diff = Math.abs(moment(diffMessage.createdAt).startOf('day').diff(moment(currentMessage.createdAt).startOf('day'), 'days'));
-    } else {
-      diff = 1;
+  getInnerComponentProps() {
+    const {containerStyle, ...props} = this.props;
+    return {
+      ...props,
+      isSameUser,
+      isSameDay
     }
-    if (diff === 0) {
-      return true;
-    }
-    return false;
-  }
-
-  isSameUser(currentMessage = {}, diffMessage = {}) {
-    if (diffMessage.user && currentMessage.user) {
-      if (diffMessage.user._id === currentMessage.user._id) {
-        return true;
-      }
-    }
-    return false;
   }
 
   renderDay() {
     if (this.props.currentMessage.createdAt) {
-      const {containerStyle, ...other} = this.props;
-      const dayProps = {
-        ...other,
-        isSameUser: this.isSameUser,
-        isSameDay: this.isSameDay,
-      };
+      const dayProps = this.getInnerComponentProps();
       if (this.props.renderDay) {
         return this.props.renderDay(dayProps);
       }
@@ -51,43 +36,44 @@ export default class Message extends React.Component {
   }
 
   renderBubble() {
-    const {containerStyle, ...other} = this.props;
-    const bubbleProps = {
-      ...other,
-      isSameUser: this.isSameUser,
-      isSameDay: this.isSameDay,
-    };
+    const bubbleProps = this.getInnerComponentProps();
     if (this.props.renderBubble) {
       return this.props.renderBubble(bubbleProps);
     }
     return <Bubble {...bubbleProps}/>;
   }
 
-  renderAvatar() {
-    if (this.props.user._id !== this.props.currentMessage.user._id) {
-      const {containerStyle, ...other} = this.props;
-      const avatarProps = {
-        ...other,
-        isSameUser: this.isSameUser,
-        isSameDay: this.isSameDay,
-      };
-
-      return <Avatar {...avatarProps}/>;
+  renderSystemMessage() {
+    const systemMessageProps = this.getInnerComponentProps();
+    if (this.props.renderSystemMessage) {
+      return this.props.renderSystemMessage(systemMessageProps);
     }
-    return null;
+    return <SystemMessage {...systemMessageProps} />;
+  }
+
+  renderAvatar() {
+    if (this.props.user._id === this.props.currentMessage.user._id && !this.props.showUserAvatar) {
+      return null;
+    }
+    const avatarProps = this.getInnerComponentProps();
+    const { currentMessage } = avatarProps;
+    if (currentMessage.user.avatar === null) {
+      return null;
+    }
+    return <Avatar {...avatarProps} />;
   }
 
   render() {
     return (
       <View>
         {this.renderDay()}
-        <View style={[styles[this.props.position].container, {
-          marginBottom: this.isSameUser(this.props.currentMessage, this.props.nextMessage) ? 2 : 10,
-        }, this.props.containerStyle[this.props.position]]}>
-          {this.props.position === 'left' ? this.renderAvatar() : null}
-          {this.renderBubble()}
-          {this.props.position === 'right' ? this.renderAvatar() : null}
-        </View>
+        {this.props.currentMessage.system ? 
+          this.renderSystemMessage() :
+          <View style={[styles[this.props.position].container, { marginBottom: isSameUser(this.props.currentMessage, this.props.nextMessage) ? 2 : 10 }, this.props.containerStyle[this.props.position]]}>
+            {this.props.position === "left" ? this.renderAvatar() : null}
+            {this.renderBubble()}
+            {this.props.position === "right" ? this.renderAvatar() : null}
+          </View>}
       </View>
     );
   }
@@ -115,9 +101,10 @@ const styles = {
 };
 
 Message.defaultProps = {
-  renderAvatar: null,
+  renderAvatar: undefined,
   renderBubble: null,
   renderDay: null,
+  renderSystemMessage: null,
   position: 'left',
   currentMessage: {},
   nextMessage: {},
@@ -127,16 +114,18 @@ Message.defaultProps = {
 };
 
 Message.propTypes = {
-  renderAvatar: React.PropTypes.func,
-  renderBubble: React.PropTypes.func,
-  renderDay: React.PropTypes.func,
-  position: React.PropTypes.oneOf(['left', 'right']),
-  currentMessage: React.PropTypes.object,
-  nextMessage: React.PropTypes.object,
-  previousMessage: React.PropTypes.object,
-  user: React.PropTypes.object,
-  containerStyle: React.PropTypes.shape({
-    left: View.propTypes.style,
-    right: View.propTypes.style,
+  renderAvatar: PropTypes.func,
+  showUserAvatar: PropTypes.bool,
+  renderBubble: PropTypes.func,
+  renderDay: PropTypes.func,
+  renderSystemMessage: PropTypes.func,
+  position: PropTypes.oneOf(['left', 'right']),
+  currentMessage: PropTypes.object,
+  nextMessage: PropTypes.object,
+  previousMessage: PropTypes.object,
+  user: PropTypes.object,
+  containerStyle: PropTypes.shape({
+    left: ViewPropTypes.style,
+    right: ViewPropTypes.style,
   }),
 };
