@@ -1,8 +1,10 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 
 import {
   ListView,
   View,
+  StyleSheet,
 } from 'react-native';
 
 import shallowequal from 'shallowequal';
@@ -105,7 +107,9 @@ export default class MessageContainer extends React.Component {
       console.warn('GiftedChat: `_id` is missing for message', JSON.stringify(message));
     }
     if (!message.user) {
-      console.warn('GiftedChat: `user` is missing for message', JSON.stringify(message));
+      if (!message.system) {
+        console.warn("GiftedChat: `user` is missing for message", JSON.stringify(message));
+      }
       message.user = {};
     }
 
@@ -137,13 +141,17 @@ export default class MessageContainer extends React.Component {
 
   render() {
     return (
-      <View ref='container' style={{flex:1}}>
+      <View
+        ref='container'
+        style={styles.container}
+      >
         <ListView
           enableEmptySections={true}
-          keyboardShouldPersistTaps={true}
           automaticallyAdjustContentInsets={false}
           initialListSize={20}
           pageSize={20}
+
+          {...this.props.listViewProps}
 
           dataSource={this.state.dataSource}
 
@@ -157,6 +165,12 @@ export default class MessageContainer extends React.Component {
   }
 }
 
+const styles = StyleSheet.create({
+  container: {
+    flex: 1
+  }
+});
+
 MessageContainer.defaultProps = {
   messages: [],
   user: {},
@@ -167,9 +181,10 @@ MessageContainer.defaultProps = {
 };
 
 MessageContainer.propTypes = {
-  messages: React.PropTypes.array,
-  user: React.PropTypes.object,
-  renderFooter: React.PropTypes.func,
-  renderMessage: React.PropTypes.func,
-  onLoadEarlier: React.PropTypes.func,
+  messages: PropTypes.array,
+  user: PropTypes.object,
+  renderFooter: PropTypes.func,
+  renderMessage: PropTypes.func,
+  onLoadEarlier: PropTypes.func,
+  listViewProps: PropTypes.object,
 };
